XWIKI-23043: Provide validation and UI for group of macro parameters belonging to same feature
  * Move java API to xwiki-platform-wysiwyg-api module
