+++ xwiki-platform-core/xwiki-platform-rendering/xwiki-platform-rendering-xwiki/src/main/resources/META-INF/components.txt
@@ -3,7 +3,6 @@
 500:org.xwiki.rendering.internal.renderer.XWikiPageAttachmentURILabelGenerator
 500:org.xwiki.rendering.internal.util.XWikiIconProvider
 500:org.xwiki.rendering.internal.util.XWikiErrorBlockGenerator
-org.xwiki.rendering.internal.util.ui.MacroDescriptorUIFactory
 500:org.xwiki.rendering.internal.wiki.XWikiWikiModel
 500:org.xwiki.rendering.internal.macro.XWikiHTMLRawBlockFilter
 org.xwiki.rendering.internal.macro.ContentMacroRequiredRightsAnalyzer
