+++ xwiki-platform-core/xwiki-platform-wysiwyg/xwiki-platform-wysiwyg-api/src/main/java/org/xwiki/wysiwyg/macro/AbstractMacroUINodeType.java
@@ -17,7 +17,9 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.xwiki.rendering.internal.util.ui;
+package org.xwiki.wysiwyg.macro;
+
+import org.xwiki.stability.Unstable;
 
 /**
  * Possible types of nodes.
@@ -25,6 +27,7 @@ package org.xwiki.rendering.internal.util.ui;
  * @version $Id$
  * @since 17.5.0RC1
  */
+@Unstable
 public enum AbstractMacroUINodeType
 {
     /**
