+++ xwiki-platform-core/xwiki-platform-wysiwyg/xwiki-platform-wysiwyg-api/src/main/java/org/xwiki/wysiwyg/script/WysiwygEditorScriptService.java
@@ -19,7 +19,9 @@
  */
 package org.xwiki.wysiwyg.script;
 
+import java.io.StringReader;
 import java.util.Collection;
+import java.util.LinkedHashMap;
 import java.util.Map;
 
 import javax.inject.Inject;
@@ -35,6 +37,15 @@ import org.xwiki.model.reference.AttachmentReference;
 import org.xwiki.model.reference.DocumentReference;
 import org.xwiki.model.reference.EntityReference;
 import org.xwiki.model.reference.EntityReferenceSerializer;
+import org.xwiki.rendering.block.Block;
+import org.xwiki.rendering.block.MacroBlock;
+import org.xwiki.rendering.block.XDOM;
+import org.xwiki.rendering.block.match.MacroBlockMatcher;
+import org.xwiki.rendering.macro.MacroId;
+import org.xwiki.rendering.macro.MacroIdFactory;
+import org.xwiki.rendering.macro.MacroLookupException;
+import org.xwiki.rendering.macro.MacroManager;
+import org.xwiki.rendering.macro.descriptor.MacroDescriptor;
 import org.xwiki.rendering.parser.ParseException;
 import org.xwiki.rendering.parser.Parser;
 import org.xwiki.rendering.renderer.PrintRendererFactory;
@@ -42,9 +53,12 @@ import org.xwiki.rendering.syntax.Syntax;
 import org.xwiki.script.service.ScriptService;
 import org.xwiki.security.authorization.ContextualAuthorizationManager;
 import org.xwiki.security.authorization.Right;
+import org.xwiki.stability.Unstable;
 import org.xwiki.store.TemporaryAttachmentSessionsManager;
 import org.xwiki.wysiwyg.converter.HTMLConverter;
 import org.xwiki.wysiwyg.importer.AttachmentImporter;
+import org.xwiki.wysiwyg.macro.MacroDescriptorUI;
+import org.xwiki.wysiwyg.internal.macro.MacroDescriptorUIFactory;
 
 import com.xpn.xwiki.XWikiContext;
 import com.xpn.xwiki.XWikiException;
@@ -59,6 +73,7 @@ import com.xpn.xwiki.doc.XWikiDocument;
 @Component
 @Named("wysiwyg")
 @Singleton
+@SuppressWarnings("checkstyle:ClassFanOutComplexity")
 public class WysiwygEditorScriptService implements ScriptService
 {
     /**
@@ -100,6 +115,19 @@ public class WysiwygEditorScriptService implements ScriptService
 
     @Inject
     private TemporaryAttachmentSessionsManager temporaryAttachmentSessionsManager;
+
+    @Inject
+    private MacroDescriptorUIFactory macroDescriptorUIFactory;
+
+    @Inject
+    private MacroManager macroManager;
+
+    @Inject
+    private MacroIdFactory macroIdFactory;
+
+    @Inject
+    @Named("html/5.0")
+    private Parser html5Parser;
 
     /**
      * Checks if there is a parser and a renderer available for the specified syntax.
@@ -446,6 +474,86 @@ public class WysiwygEditorScriptService implements ScriptService
             return Syntax.valueOf(syntaxId);
         } catch (ParseException e) {
             throw new RuntimeException(String.format("Invalid syntax [%s]", syntaxId), e);
+        }
+    }
+
+    /**
+     * Get a macro descriptor UI information to be used for configuring a macro.
+     * @param macroIdAsString the identifier of a macro
+     * @return an instance of {@link MacroDescriptorUI} containing all info for configuring the macro or {@code null}
+     * if it couldn't be found or initialized.
+     *
+     * @since 17.5.0RC1
+     */
+    @Unstable
+    public MacroDescriptorUI getMacroDescriptorUI(String macroIdAsString)
+    {
+        try {
+            MacroId macroId = this.resolveMacroId(macroIdAsString);
+            if (macroId != null && this.macroManager.exists(macroId, true)) {
+                MacroDescriptor descriptor = this.macroManager.getMacro(macroId).getDescriptor();
+                return this.macroDescriptorUIFactory.buildMacroDescriptorUI(descriptor);
+            }
+        } catch (MacroLookupException e) {
+            this.logger.warn("Failed to lookup macro id [{}]. Root cause is: [{}]", macroIdAsString,
+                ExceptionUtils.getRootCauseMessage(e));
+        }
+        return null;
+    }
+
+    /**
+     * Extract macro parameters values from the given html fragment that is supposed to contain information
+     * representing the macro values.
+     * @param macroIdAsString the id of the macro to find in the html fragment
+     * @param html an html fragment containing information about the macro.
+     * @return a map of parameters also containing information about the content with the key {@code $content}.
+     * @since 17.5.0RC1
+     */
+    @Unstable
+    public Map<String, String> getMacroParametersFromHTML(String macroIdAsString, String html)
+    {
+        MacroId macroId = this.resolveMacroId(macroIdAsString);
+        Map<String, String> result = new LinkedHashMap<>();
+        if (macroId != null) {
+            String htmlToParse = String.format("<html><body>%s</body></html>", html);
+            XDOM xdom = null;
+            try {
+                xdom = this.html5Parser.parse(new StringReader(htmlToParse));
+                extractMacroParameters(xdom, macroId, result);
+            } catch (ParseException e) {
+                this.logger.error("Error while parsing html for macro [{}]: [{}]", macroIdAsString, htmlToParse, e);
+            }
+        }
+
+        return result;
+    }
+
+    private static void extractMacroParameters(XDOM xdom, MacroId macroId, Map<String, String> result)
+    {
+        if (xdom != null) {
+            Block block = xdom.getFirstBlock(new MacroBlockMatcher(macroId.getId()), Block.Axes.DESCENDANT);
+            if (block instanceof MacroBlock macroBlock) {
+                result.putAll(macroBlock.getParameters());
+                if (macroBlock.getContent() != null) {
+                    result.put("$content", macroBlock.getContent());
+                }
+            }
+        }
+    }
+
+    /**
+     * @param macroIdAsString a string representing a macro id
+     * @return the resolved macro id or {@code null} if resolving the given string fails
+     * @since 10.10RC1
+     */
+    private MacroId resolveMacroId(String macroIdAsString)
+    {
+        try {
+            return this.macroIdFactory.createMacroId(macroIdAsString);
+        } catch (ParseException e) {
+            this.logger.warn("Failed to resolve macro id [{}]. Root cause is: [{}]", macroIdAsString,
+                ExceptionUtils.getRootCauseMessage(e));
+            return null;
         }
     }
 }
