+++ xwiki-platform-core/xwiki-platform-wysiwyg/xwiki-platform-wysiwyg-api/src/main/resources/META-INF/components.txt
@@ -5,9 +5,10 @@ org.xwiki.wysiwyg.internal.cleaner.LineBreakFilter
 org.xwiki.wysiwyg.internal.cleaner.NestedAnchorsFilter
 org.xwiki.wysiwyg.internal.cleaner.StandAloneMacroFilter
 org.xwiki.wysiwyg.internal.converter.DefaultHTMLConverter
+org.xwiki.wysiwyg.internal.converter.DefaultRequestParameterConverter
+org.xwiki.wysiwyg.internal.converter.HTMLRequestParameterConverter
 org.xwiki.wysiwyg.internal.filter.http.MutableHttpServletRequestFactory
 org.xwiki.wysiwyg.internal.importer.OfficeAttachmentImporter
 org.xwiki.wysiwyg.internal.importer.OfficeMacroImporter
+org.xwiki.wysiwyg.internal.macro.MacroDescriptorUIFactory
 org.xwiki.wysiwyg.script.WysiwygEditorScriptService
-org.xwiki.wysiwyg.internal.converter.DefaultRequestParameterConverter
-org.xwiki.wysiwyg.internal.converter.HTMLRequestParameterConverter
